Modify : Comment Dto Long으로 변경

+++ server/src/main/java/com/server/soopool/comment/dto/CommentDeleteDto.java
@@ -4,5 +4,5 @@ import lombok.Getter;
 
 @Getter
 public class CommentDeleteDto {
-    private Integer groupNumber;
+    private Long groupNumber;
 }
